Implement Docker tag tracking (#28)
This implements a field `dockerImage.trackMutableTag` similar to `gitConfig.trackMutableRef`.

The assumption is that when building from a particular branch or for a particular release track, every built version will be stamped with two tags: an "immutable" tag like `main---000123-abcd0123` and a "mutable" tag like `main`. The actual tag that will be put in the container spec will be the immutable tag; the automation will ensure that the tag field is updated to an immutable tag pointing at the same image that the mutable tag currently does.

This implementation assumes you are using Google's Artifact Registry and that all images that use trackMutableRef are stored in a single AR repository (ie, set of Docker image repositories).